move to Utility.cs
